# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "paopao-ce API",
    "description": "# paopao-ce API 文档\n\n欢迎使用 paopao-ce API 文档！\n\npaopao-ce 是一个开源的泡泡社区系统，提供完整的社交媒体功能。\n\n## 鉴权说明\n\n大部分API需要JWT Token进行鉴权，请在请求头中包含：\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## 响应格式\n\n所有API响应都遵循以下格式：\n```json\n{\n  \"code\": 0,\n  \"message\": \"success\",\n  \"data\": {}\n}\n```\n\n## 错误码说明\n\n- `0`: 成功\n- `10001`: 未授权\n- `10002`: 参数错误\n- `10003`: 权限不足\n- `10004`: 资源不存在\n\n## 分页参数\n\n支持分页的API使用以下查询参数：\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)\n",
    "version": "1.0.0",
    "contact": {
      "name": "paopao-ce",
      "url": "https://github.com/rocboss/paopao-ce"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8008",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.paopao.info",
      "description": "生产环境"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/user/info": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取用户信息",
        "description": "获取指定用户的基本信息，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 可以获取自己的信息或他人的公开信息",
        "operationId": "getUserInfo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "用户名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
    "/user/messages": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取消息列表",
        "description": "获取当前用户的消息列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `style`: 消息类型 (1:系统消息, 2:私信, 3:@提醒)\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "getMessages",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "style",
            "in": "query",
            "description": "消息类型",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [1, 2, 3]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMessagesResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/messages/stream": {
      "get": {
        "tags": ["用户管理"],
        "summary": "SSE获取消息列表",
        "description": "建立Server-Sent Events连接，实时获取用户的未读消息通知，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**连接特性：**\n- Server-Sent Events (SSE) 流式响应\n- 每5秒自动推送新消息\n- 只推送未读消息，避免重复通知\n- 自动断开连接处理\n\n**SSE事件类型：**\n- `connected`: 连接建立确认\n- `message`: 新消息通知\n- `error`: 错误信息\n\n**消息数据格式：**\n```json\n{\n  \"id\": 123,\n  \"type\": 1,\n  \"brief\": \"消息摘要\",\n  \"content\": \"消息内容\",\n  \"sender_user_id\": 456,\n  \"receiver_user_id\": 789,\n  \"post_id\": 101,\n  \"comment_id\": 202,\n  \"reply_id\": 303,\n  \"is_read\": 0,\n  \"created_on\": 1640995200\n}\n```",
        "operationId": "streamMessages",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE流连接成功",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events 流数据"
                },
                "example": "event: connected\ndata: {\"user_id\": 123, \"timestamp\": 1640995200}\n\nevent: message\ndata: {\"id\": 456, \"type\": 1, \"brief\": \"新消息\", \"content\": \"消息内容\", \"sender_user_id\": 789, \"is_read\": 0, \"created_on\": 1640995200}\n\n"
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/user/message/read": {
      "post": {
        "tags": ["用户管理"],
        "summary": "标记消息已读",
        "description": "标记指定消息为已读状态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 只能标记自己的消息",
        "operationId": "readMessage",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadMessageReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "消息不存在"
          }
        }
      }
    },
    "/user/message/readall": {
      "post": {
        "tags": ["用户管理"],
        "summary": "标记所有消息已读",
        "description": "标记当前用户的所有未读消息为已读，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token",
        "operationId": "readAllMessage",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/whisper": {
      "post": {
        "tags": ["用户管理"],
        "summary": "发送私信",
        "description": "向指定用户发送私信，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 不能给自己发私信\n- 内容不能为空",
        "operationId": "sendUserWhisper",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendWhisperReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/collections": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取收藏列表",
        "description": "获取当前用户的收藏列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "getCollections",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCollectionsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/stars": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取点赞列表",
        "description": "获取当前用户的点赞列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "getStars",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetStarsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/phone": {
      "post": {
        "tags": ["用户管理"],
        "summary": "绑定手机号",
        "description": "绑定手机号到当前用户账户，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 需要先获取短信验证码\n- 手机号不能已被其他用户绑定",
        "operationId": "userPhoneBind",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPhoneBindReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "绑定成功"
          },
          "400": {
            "description": "参数错误或手机号已被绑定"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/password": {
      "post": {
        "tags": ["用户管理"],
        "summary": "修改密码",
        "description": "修改当前用户的登录密码，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 需要提供旧密码进行验证\n- 新密码不能为空",
        "operationId": "changePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "参数错误或旧密码不正确"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/nickname": {
      "post": {
        "tags": ["用户管理"],
        "summary": "修改昵称",
        "description": "修改当前用户的昵称，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 昵称不能为空\n- 昵称长度限制在2-20个字符",
        "operationId": "changeNickname",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeNicknameReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/avatar": {
      "post": {
        "tags": ["用户管理"],
        "summary": "修改头像",
        "description": "修改当前用户的头像，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- avatar字段为头像URL地址",
        "operationId": "changeAvatar",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAvatarReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/suggest/users": {
      "get": {
        "tags": ["用户管理"],
        "summary": "搜索用户",
        "description": "根据关键词搜索用户，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `keyword`: 搜索关键词",
        "operationId": "suggestUsers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestUsersResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/suggest/tags": {
      "get": {
        "tags": ["用户管理"],
        "summary": "搜索标签",
        "description": "根据关键词搜索标签，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `keyword`: 搜索关键词",
        "operationId": "suggestTags",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestTagsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/post/star": {
      "get": {
        "tags": ["动态管理"],
        "summary": "获取动态点赞状态",
        "description": "获取指定动态的点赞状态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 动态ID",
        "operationId": "tweetStarStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "动态ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TweetStarStatusResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": ["动态管理"],
        "summary": "点赞动态",
        "description": "对指定动态进行点赞操作，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 重复点赞会取消点赞",
        "operationId": "starTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StarTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StarTweetResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "动态不存在"
          }
        }
      }
    },
    "/post/collection": {
      "get": {
        "tags": ["动态管理"],
        "summary": "获取动态收藏状态",
        "description": "获取指定动态的收藏状态，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**请求参数：**\n- `id`: 动态ID",
        "operationId": "tweetCollectionStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "动态ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TweetCollectionStatusResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": ["动态管理"],
        "summary": "收藏动态",
        "description": "对指定动态进行收藏操作，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 重复收藏会取消收藏",
        "operationId": "collectionTweet",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionTweetReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionTweetResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "动态不存在"
          }
        }
      }
    },
    "/user/articles/collections": {
      "get": {
        "tags": ["长文章管理"],
        "summary": "获取长文章收藏列表",
        "description": "获取当前用户的长文章收藏列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "getUserArticleCollections",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserArticleCollectionsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/user/articles/stars": {
      "get": {
        "tags": ["长文章管理"],
        "summary": "获取长文章点赞列表",
        "description": "获取当前用户的长文章点赞列表，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**分页参数：**\n- `page`: 页码 (默认: 1)\n- `page_size`: 每页数量 (默认: 20)",
        "operationId": "getUserArticleStars",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserArticleStarsResp"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/sync/index": {
      "get": {
        "tags": ["用户管理"],
        "summary": "同步搜索索引",
        "description": "同步搜索索引数据，需要JWT鉴权\n\n**鉴权方式：**\n- Bearer Token\n\n**注意事项：**\n- 通常用于后台管理或定时任务",
        "operationId": "syncSearchIndex",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "同步成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserInfoResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户ID"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "status": {
            "type": "integer",
            "description": "用户状态"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "balance": {
            "type": "integer",
            "description": "账户余额"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "is_admin": {
            "type": "boolean",
            "description": "是否管理员"
          },
          "created_on": {
            "type": "integer",
            "description": "创建时间戳"
          },
          "follows": {
            "type": "integer",
            "description": "关注数"
          },
          "followings": {
            "type": "integer",
            "description": "粉丝数"
          },
          "tweets_count": {
            "type": "integer",
            "description": "动态数量"
          },
          "experience": {
            "type": "integer",
            "description": "经验值"
          },
          "level": {
            "type": "integer",
            "description": "等级"
          }
        }
      },
      "GetMessagesResp": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "total": {
            "type": "integer",
            "description": "总数"
          },
          "page": {
            "type": "integer",
            "description": "当前页码"
          },
          "page_size": {
            "type": "integer",
            "description": "每页数量"
          }
        }
      },
...
      "ReadMessageReq": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "消息ID"
          }
        }
      },
...
    }
  }
}